Disable las file download for selected rows #2602
